pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sopht-mpi"
version = "0.0.1"
description = "Scalable One-stop Platform for Hydroelastic Things (SOPHT) MPI solver."
readme = "README.md"
authors = ["Fan Kiat Chan <fankiat@hotmail.com>", "Yashraj Bhosale <bhosale2@illinois.edu>"]
license = "MIT"
repository = "https://github.com/fankiat/sopht-mpi"

# ...
packages = [
    { include = "sopht_mpi" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
sopht_backend = {git = "https://github.com/bhosale2/sopht-backend.git", rev = "main"}
mpi4py = "^3.1"
# add mpi-fft here later if pip works out...

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.19.0"
autoflake8 = "^0.4"
pytest = "^7.1"
pytest-mpi = "^0.6"

[tool.black]
# https://github.com/psf/black
target-version = ["py38"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
